kba-UID2-873-node-application-logging: refactor logging code into a helper
